cmake: emu: renode: Redirect simulated console to stdout
The renode simulator integration with Twister did not pass the
console output of the simulated device to stdout, meaning that
the test runner was not able to parse it for test output.

If a UART console has been defined, connect it to stdout of the
renode process.

This enables the use of west twister --simulation renode.

Signed-off-by: Aksel Skauge Mellbye <aksel.mellbye@silabs.com>